Brazilian Portuguese translation updated by Luiz Armesto.
2007-12-22  Leonardo Ferreira Fontenelle  <leonardof@svn.gnome.org>

	* pt_BR.po: Brazilian Portuguese translation updated by Luiz Armesto.

svn path=/trunk/; revision=411

original bzr status:
revno: 349
author: Leonardo Ferreira Fontenelle <leonardof@svn.gnome.org>
committer: Leonardo Ferreira Fontenelle <leonardof@src.gnome.org>
timestamp: Sat 2007-12-22 23:41:35 +0000

+++ src/view.py
@@ -29,7 +29,7 @@ from config import settings
 
 
 def get_seconds_remaining_in_day():
-    return 86400 - ((time.time()-time.timezone-time.daylight*(60*60)) % 86400) + 4
+    return int(86400 - ((time.time()-time.timezone-time.daylight*(60*60)) % 86400) + 4)
 
 
 class ActivityView(gtk.VBox):
